+++ native/golang/job.go
@@ -18,10 +18,13 @@ import (
 	"os"
 	"strconv"
 	"strings"
+	t "zowe-native-proto/ioserver/types/common"
+	"zowe-native-proto/ioserver/types/jobs"
+	utils "zowe-native-proto/ioserver/utils"
 )
 
 func HandleListJobsRequest(jsonData []byte) {
-	var listRequest ListJobsRequest
+	var listRequest jobs.ListJobsRequest
 	err := json.Unmarshal(jsonData, &listRequest)
 	if err != nil {
 		// log.Println("Error decoding ListDatasetsRequest:", err)
@@ -33,24 +36,24 @@ func HandleListJobsRequest(jsonData []byte) {
 		args = append(args, "--owner", listRequest.Owner)
 	}
 
-	out, err := buildCommand(args).Output()
+	out, err := utils.BuildCommand(args).Output()
 	if err != nil {
 		log.Println("Error executing command:", err)
 		return
 	}
 
-	jobs := strings.Split(strings.TrimSpace(string(out)), "\n")
+	returnedJobs := strings.Split(strings.TrimSpace(string(out)), "\n")
 
-	jobsResponse := ListJobsResponse{
-		Items: make([]Job, len(jobs)),
+	jobsResponse := jobs.ListJobsResponse{
+		Items: make([]t.Job, len(returnedJobs)),
 	}
 
-	for i, job := range jobs {
+	for i, job := range returnedJobs {
 		vals := strings.Split(job, ",")
 		if len(vals) < 4 {
 			continue
 		}
-		jobsResponse.Items[i] = Job{
+		jobsResponse.Items[i] = t.Job{
 			Id:      vals[0],
 			Retcode: vals[1],
 			Name:    strings.TrimSpace(vals[2]),
@@ -67,7 +70,7 @@ func HandleListJobsRequest(jsonData []byte) {
 }
 
 func HandleListSpoolsRequest(jsonData []byte) {
-	var listRequest ListSpoolsRequest
+	var listRequest jobs.ListSpoolsRequest
 	err := json.Unmarshal(jsonData, &listRequest)
 	if err != nil {
 		// log.Println("Error decoding ListSpoolsRequest:", err)
@@ -76,7 +79,7 @@ func HandleListSpoolsRequest(jsonData []byte) {
 
 	args := []string{"./zowex", "job", "list-files", listRequest.JobId, "--rfc", "true"}
 
-	out, err := buildCommand(args).Output()
+	out, err := utils.BuildCommand(args).Output()
 	if err != nil {
 		log.Println("Error executing command:", err)
 		return
@@ -84,8 +87,8 @@ func HandleListSpoolsRequest(jsonData []byte) {
 
 	spools := strings.Split(strings.TrimSpace(string(out)), "\n")
 
-	response := ListSpoolsResponse{
-		Items: make([]Spool, len(spools)),
+	response := jobs.ListSpoolsResponse{
+		Items: make([]t.Spool, len(spools)),
 	}
 
 	for i, spool := range spools {
@@ -97,7 +100,7 @@ func HandleListSpoolsRequest(jsonData []byte) {
 		if err != nil {
 			continue
 		}
-		response.Items[i] = Spool{
+		response.Items[i] = t.Spool{
 			Id:       id,
 			DdName:   vals[0],
 			StepName: vals[3],
@@ -115,7 +118,7 @@ func HandleListSpoolsRequest(jsonData []byte) {
 }
 
 func HandleReadSpoolRequest(jsonData []byte) {
-	var request ReadSpoolRequest
+	var request jobs.ReadSpoolRequest
 	err := json.Unmarshal(jsonData, &request)
 	if err != nil {
 		// log.Println("Error decoding ReadSpoolRequest:", err)
@@ -127,15 +130,15 @@ func HandleReadSpoolRequest(jsonData []byte) {
 	if hasEncoding {
 		args = append(args, "--encoding", request.Encoding, "--rfb", "true")
 	}
-	out, err := buildCommand(args).Output()
+	out, err := utils.BuildCommand(args).Output()
 	if err != nil {
 		log.Println("Error executing command:", err)
 		return
 	}
 
-	data := collectContentsAsBytes(string(out), hasEncoding)
+	data := utils.CollectContentsAsBytes(string(out), hasEncoding)
 
-	response := ReadSpoolResponse{
+	response := jobs.ReadSpoolResponse{
 		Encoding: request.Encoding,
 		DsnKey:   request.DsnKey,
 		JobId:    request.JobId,
@@ -150,7 +153,7 @@ func HandleReadSpoolRequest(jsonData []byte) {
 }
 
 func HandleGetJclRequest(jsonData []byte) {
-	var request GetJclRequest
+	var request jobs.GetJclRequest
 	err := json.Unmarshal(jsonData, &request)
 	if err != nil {
 		// log.Println("Error decoding GetJclRequest:", err)
@@ -158,13 +161,13 @@ func HandleGetJclRequest(jsonData []byte) {
 	}
 	// log.Println("GetJclRequest received:", ...)
 	args := []string{"./zowex", "job", "view-jcl", request.JobId}
-	out, err := buildCommand(args).Output()
+	out, err := utils.BuildCommand(args).Output()
 	if err != nil {
 		log.Println("Error executing command:", err)
 		return
 	}
 
-	response := GetJclResponse{
+	response := jobs.GetJclResponse{
 		JobId: request.JobId,
 		Data:  string(out),
 	}
@@ -177,30 +180,30 @@ func HandleGetJclRequest(jsonData []byte) {
 }
 
 func HandleGetStatusRequest(jsonData []byte) {
-	var request GetJclRequest
+	var request jobs.GetJclRequest
 	err := json.Unmarshal(jsonData, &request)
 	if err != nil {
 		return
 	}
 	args := []string{"./zowex", "job", "view-status", request.JobId, "--rfc", "true"}
-	out, err := buildCommand(args).Output()
+	out, err := utils.BuildCommand(args).Output()
 	if err != nil {
 		log.Println("Error executing command:", err)
 		return
 	}
-	jobs := strings.Split(strings.TrimSpace(string(out)), "\n")
+	returnedJobs := strings.Split(strings.TrimSpace(string(out)), "\n")
 
 	// log.Println(jobs)
-	jobsResponse := ListJobsResponse{
-		Items: make([]Job, len(jobs)),
+	jobsResponse := jobs.ListJobsResponse{
+		Items: make([]t.Job, len(returnedJobs)),
 	}
 
-	for i, job := range jobs {
+	for i, job := range returnedJobs {
 		vals := strings.Split(job, ",")
 		if len(vals) < 4 {
 			continue
 		}
-		jobsResponse.Items[i] = Job{
+		jobsResponse.Items[i] = t.Job{
 			Id:      vals[0],
 			Retcode: vals[1],
 			Name:    strings.TrimSpace(vals[2]),