+++ packages/cli/src/view/job-file/JobFile.handler.ts
@@ -10,13 +10,13 @@
  */
 
 import type { IHandlerParameters } from "@zowe/imperative";
-import { type ReadSpool, type ZSshClient, ZSshUtils } from "zowe-native-proto-sdk";
+import { type jobs, type ZSshClient, ZSshUtils } from "zowe-native-proto-sdk";
 import { SshBaseHandler } from "../../SshBaseHandler";
 
 export default class ViewJobJclHandler extends SshBaseHandler {
-    public async processWithClient(params: IHandlerParameters, client: ZSshClient): Promise<ReadSpool.Response> {
+    public async processWithClient(params: IHandlerParameters, client: ZSshClient): Promise<jobs.ReadSpoolResponse> {
         const response = await client.jobs.readSpool({
-            dsnKey: params.arguments.dsnKey,
+            spoolId: params.arguments.dsnKey,
             jobId: params.arguments.jobId,
             encoding: params.arguments.encoding,
         });